+++ src/providers/index.ts
@@ -2,25 +2,27 @@ import { EnvProvider } from '../interfaces';
 import { AppSettingsProvider } from './app-settings.provider';
 import { SecretsProvider } from './secrets.provider';
 import { PackageJsonProvider } from './package-json.provider';
+import { LocalProvider } from './local.provider';
 
 const IntegratedProviders: Record<string, EnvProvider<any, any>> = {
     [PackageJsonProvider.key]: PackageJsonProvider,
     [AppSettingsProvider.key]: AppSettingsProvider,
-    [SecretsProvider.key]: SecretsProvider
+    [SecretsProvider.key]: SecretsProvider,
+    [LocalProvider.key]: LocalProvider
 };
 
 const IntegratedProviderConfig = [
     {
-        path: PackageJsonProvider.key,
-        type: 'integrated'
+        path: PackageJsonProvider.key
     },
     {
-        path: AppSettingsProvider.key,
-        type: 'integrated'
+        path: AppSettingsProvider.key
     },
     {
-        path: SecretsProvider.key,
-        type: 'integrated'
+        path: SecretsProvider.key
+    },
+    {
+        path: LocalProvider.key
     }
 ];